+++ app/javascript/components/index.js
@@ -1,7 +1,6 @@
 // javascript/components/index.js
 export { default as AnimatedTitle } from './AnimatedTitle'
-export { default as Groceries } from './Groceries'
+export { default as Pantry } from './Pantry'
 export { default as GroceryItem } from './GroceryItem'
-export { default as SectionModal } from './SectionModal'
-export { default as ItemModal } from './ItemModal'
+
 // Add more components here
