+++ app/views/groceries/index.html.erb
@@ -1,3 +1,3 @@
-<%= react_component("Groceries", {
-  groceryItems: @grouped_groceries
+<%= react_component("Pantry", {
+  groceries: @grouped_groceries
 }) %>
